+++ imsdk/src/main/java/com/masonsoft/imsdk/core/processor/SendMessageTypeImageValidateProcessor.java
@@ -27,7 +27,7 @@ public class SendMessageTypeImageValidateProcessor extends SendMessageTypeValida
     }
 
     @Override
-    protected boolean doTypeProcess(@NonNull IMSessionMessage target, long type) {
+    protected boolean doTypeProcess(@NonNull IMSessionMessage target, int type) {
         final StateProp<String> body = target.getIMMessage().body;
         if (body.isUnset()) {
             target.getEnqueueCallback().onEnqueueFail(
